+++ pytorch_lightning/loggers/neptune.py
@@ -20,11 +20,11 @@ import torch
 from torch import is_tensor
 
 from pytorch_lightning import _logger as log
-from pytorch_lightning.loggers.base import LightningLoggerBase
+from pytorch_lightning.loggers.base import LightningLogger
 from pytorch_lightning.utilities import rank_zero_only
 
 
-class NeptuneLogger(LightningLoggerBase):
+class NeptuneLogger(LightningLogger):
     r"""
     Log using `Neptune <https://neptune.ai>`_. Install it with pip: